feat: error package, i18n and vee-validate
affects: @platyplus/authentication-service, @platyplus/platyplus-service, @platyplus/platyplus,
@platyplus/errors, @platyplus/hasura-apollo-client

Created a distinct @platyplus/errors module used both in frontend and backend. It captures the type
of error and stores them in the vuex store.
Harmonisation of the internationalisation, with a Vue
module ready to be transferred into a dedicated package (except the handler to get the user's
locale.
Connection of Vee-validate to the new errors system - see the Signin component